refactor: remove cycle from EmployeeList
